Make all existing tests run
All existing tests can now be executed by running npm test. Mocha was
installed as a devDependency and is used by the npm test script.

The IO type depends on ramda and because of this ramda was installed
as a dependency. This could later be removed if the purpose is to make
this package dependency free.

A .gitignore file was created which includes the node_modules folder.