Make Future.ap run in parallel
This change to `ap` for Futures does not wait for its own fork to be
resolved before forking the input furture. The effect is, among other
things, that `lift` can be used in order to call a method with the
result of several futures that have been resolved in parallel.