Add Future.Memoize
Memoize wraps a `Future` and returns a new `Future` that guarantees
that the inner `Future` will only be forked once. Later forkings will
callback with the cached resolved value or error. If the memoized
`Future` is forked several times before it has been resolved, all
callbacks will be fired once the asynchronous operation is completed.