Fixed the way Ramda is required
This commit comes from my need to use Maybe only as I do not need the rest. Thus,
I do `require('ramda-fantasy/src/Maybe')`. But the problem with the old way of doing require in `ramda-fantasy` is that tree shaking can not be performed.
Although only three functions are used in this module from `ramda`, the whole library (`ramda`) is included in my build through `ramda-fantasy`.

I use rollup in order to build my code as it can perform tree shaking. Thus, I features that are not required in my code are not included in build. However, dependencies
also need to include the only parts they need.